// apps/smart-forms-app/src/functions/populate-functions/launchContext.interface.ts
export interface LaunchContext extends Extension {
  url: 'http://hl7.org/fhir/uv/sdc/StructureDefinition/sdc-questionnaire-launchContext';
  extension: LaunchContextExtensions;
}

type LaunchContextExtensions =
  | [LaunchContextName, LaunchContextType]
  | [LaunchContextName, LaunchContextType, LaunchContextDescription];

type LaunchContextName = LaunchContextNameId | LaunchContextNameCoding;
// ...
interface LaunchContextType extends Extension {
  url: 'type';
  valueCode: FhirResourceString;
}

interface LaunchContextDescription extends Extension {
  url: 'description';
  valueString: string;
}